move expression validator
