+++ dataplane-webserver/src/metrics/mod.rs
@@ -1,1 +1,2 @@
 pub mod types;
+pub mod expression_validator;
